// src/app/interfaces/Shot.ts
export interface Shot {
  'matchid': number;
  'setsequencenumber': number;
  'gamesequencenumber': number;
  'pointsequencenumber': number;
  'shotindex': number;
  'reverseshotindex': number;
  'hittingplayer': number;
  'isvolley': boolean;
  'strokeside': string;
  'p1Startx': number;
  'p1Starty': number;
  'p2Startx': number;
  'p2Starty': number;
  'ballx': number;
  'bally': number;
  'p1Bouncex': number;
  'p1Bouncey': number;
  'p2Bouncex': number;
  'p2Bouncey': number;
  'p1Endx': number;
  'p1Endy': number;
  'p2Endx': number;
  'p2Endy': number;
  'starttimestamp': number;
  'bouncetimestamp': number;
  'endtimestamp': number;
  'winningPlayer': number;
  'p1DepthCode': number;
  'p1LRCode': string;
  'p2DepthCode': number;
  'p2LRCode': string;
  'serveNumber': number;
  'serveSide': string;
  'outcome': string;
  'isNextShotVolley': boolean;
  'id': string;
  'p1WinningPercentage': number;
  'standardSituationCode': string;
  'standardSortableOutcome': string;
}